requirements.txt
# NRL Tracker - Production Dependencies
# Python 3.9+

# Web Framework
streamlit==1.52.2

# Data Processing
pandas==2.3.3
openpyxl==3.1.5

# Document Processing
python-docx==1.2.0

# HTTP Requests
requests==2.32.5

# Face API (đăng nhập bằng khuôn mặt)
flask>=3.0
flask-cors>=4.0
Pillow>=10.0

# NOTE (Windows/Python 3.11):
# `face_recognition` phụ thuộc `dlib` và thường yêu cầu CMake + MSVC để build trên Windows.
# App `app.py` đã hỗ trợ backend `lbph` (OpenCV-only) để chạy được mà không cần dlib.
opencv-contrib-python>=4.8
pandas>=2.0
streamlit>=1.0
watchdog>=2.0

# Optional (nếu bạn đã cài được dlib/face_recognition):
# face-recognition==1.3.0


# Linting
flake8
black
isort
pre-commit

# Development
rich
pytest
pytest-cov
alive-progress

